fix #279: save detect results (#281)
* fix #279: save detect results

* rename

* set device as arg

* rm bash file
